Move recipe_tag config out of recipe
